Test the scenario with and without DjOne.
Additionally this commit slightly changes the code in tests to be more
readable and reusable. In order to be able to specify in scenario
whether or not we want to run with DjOne I am moving the plugin setup to
separate method. DjOne is no longer added to DelayedJob while loading
the suite.