Implementation for processing_id and handling failures
When the job goes into processing we want to be able to schedule another
one of the same kind, so that we always finish up with the job with the
most current state.

Also when the job fails for the last time we need to clear the unique_id
so that another instances of that job can take over.